feat(3): Update Darwin storage and locker naming, instanceId added to locks
The main purpose of this commit is to refactor code by changing all instances of 'autoupdate' to 'darwin'. Changes have also been made in SQL script file names and content. Additionally instanceId was added to leaseKey for identifying node of creation. Also method for removing all locks from one particular node was added for case when locks are kept through restart of node.